feat: conditional excel feature, CI, githooks, and clippy fixes
- Make excel feature optional behind default feature flag
- Add CI with lint, build matrix (ubuntu/macos, feature combos), integration tests
- Add .githooks/pre-commit with clippy, fmt, build, test
- Fix YAML syntax in CI workflow (heredoc -> printf)
- Resolve all clippy warnings (dead_code, module_inception, unnecessary_fallible_conversions)

+++ src/commands/listen.rs
@@ -292,7 +292,7 @@ pub async fn run(url: String, args: ListenArgs) -> Result<()> {
                             let event = NotifyEvent {
                                 channel: n.channel().to_string(),
                                 payload: n.payload().to_string(),
-                                pid: n.process_id() as i32,
+                                pid: n.process_id(),
                             };
 
                             debug!(
